refactor(api): moved API stuff to utils/API
- Also includes a major refactor to Hooks because of changes in React Query v5.
